Split events to separate file, add record example
