+++ Cargo.toml
@@ -15,18 +15,7 @@ documentation = "https://henquist.github.io/wasapi-rs/docs/0.5.0/wasapi/"
 [dependencies.windows]
 version = "0.23.0"
 features = [
-    "Data_Xml_Dom",
-    "Win32_Foundation",
-    "Win32_Security",
     "Win32_System_Threading",
-    "Win32_UI_WindowsAndMessaging",
-    "Win32_Media_Audio_CoreAudio",
-    "Win32_Media_Multimedia",
-    "Win32_Storage_StructuredStorage",
-    "Win32_System_Com",
-    "Win32_System_PropertiesSystem",
-    "Win32_System_SystemServices",
-    "Win32_System_OleAutomation",
 ]
 
 [dependencies]
